Move DOM manipulation to a different file
